data/update file remove cache folder

+++ infra.php
@@ -9,12 +9,14 @@ if (!is_file('vendor/autoload.php')) {
 require_once('vendor/autoload.php');
 
 $path = Path::$conf;
+
 if (!is_dir($path['cache'])) {
 	Config::$install = true;
 	mkdir($path['cache']);
 }
-if (is_file($path['data'].'update')) {
+if ($path['fs']&&is_file($path['data'].'update')) {
 	header('Infrajs-Path-Update:true');
+	Path::fullrmdir($path['cache']);
 	Config::$install = true;
 	unlink($path['data'].'update');
 }
